refactor: put default program into cmd/gilfoyle folder
