+++ cmd/gilfoyle/config.go
@@ -1,4 +1,4 @@
-package cmd
+package gilfoyle
 
 import (
 	"fmt"
