+++ cmd/gilfoyle/serve.go
@@ -1,13 +1,13 @@
-package cmd
+package gilfoyle
 
 import (
 	"context"
 	"fmt"
+	"github.com/dreamvo/gilfoyle"
 	"github.com/dreamvo/gilfoyle/api"
 	"github.com/dreamvo/gilfoyle/api/db"
 	"github.com/dreamvo/gilfoyle/config"
 	"github.com/dreamvo/gilfoyle/ent/migrate"
-	"github.com/dreamvo/gilfoyle/logger"
 	"github.com/gin-gonic/gin"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
@@ -28,7 +28,7 @@ var serveCmd = &cobra.Command{
 	Use:   "serve",
 	Short: "Launch the REST API",
 	Run: func(cmd *cobra.Command, args []string) {
-		logger := logger.New()
+		logger := gilfoyle.Logger
 
 		logger.Info("Initializing API server")
 		logger.Info("Environment", zap.Bool("debug", config.GetConfig().Settings.Debug))
