+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/dreamvo/gilfoyle/cmd/gilfoyle"
+
+func main() {
+	gilfoyle.Execute()
+}
