+++ go.mod
@@ -11,19 +11,19 @@ require (
 	github.com/go-playground/locales v0.13.0
 	github.com/go-playground/universal-translator v0.17.0
 	github.com/go-playground/validator/v10 v10.4.0
-	github.com/google/uuid v1.1.1
+	github.com/google/uuid v1.1.2
+	github.com/goreleaser/goreleaser v0.147.1 // indirect
 	github.com/jinzhu/configor v1.2.0
 	github.com/lib/pq v1.8.0
 	github.com/mailru/easyjson v0.7.3 // indirect
 	github.com/mattn/go-sqlite3 v1.14.0
 	github.com/minio/minio-go/v7 v7.0.5
-	github.com/spf13/cobra v1.0.0
+	github.com/spf13/cobra v1.1.1
 	github.com/stretchr/testify v1.6.1
 	github.com/swaggo/files v0.0.0-20190704085106-630677cd5c14
 	github.com/swaggo/gin-swagger v1.2.0
 	github.com/swaggo/swag v1.6.7
 	github.com/ulule/gostorages v0.2.3
 	go.uber.org/zap v1.10.0
-	golang.org/x/tools v0.0.0-20200811215021-48a8ffc5b207 // indirect
 	gopkg.in/yaml.v2 v2.3.0
 )
