--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/dreamvo/gilfoyle/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
